+++ src/compiler/sema_liveness.c
@@ -261,6 +261,7 @@ RETRY:
 		case EXPR_EMBED:
 		case EXPR_MACRO_BODY:
 		case EXPR_OTHER_CONTEXT:
+		case EXPR_MEMBER_GET:
 			UNREACHABLE
 		case EXPR_DESIGNATOR:
 			sema_trace_expr_liveness(expr->designator_expr.value);